김재령 : Service, ServiceImpl로 구성변경
